Change how terms are searched for in update_standardization.py
Terms are now saved to a dictionary the first time they are searched for with the search result so that other instances of that same format can be skipped. This saves on searching for the same term many times.

Previously, every single term was searched for and if there was no match the script checked if it had been found already.