Error handling for no format match, csv_merge.py
Didn't think this was necessary since the csv should be updated before running the script, and then made this error myself and found it confusing. So added a try/except. No match is a type error because value is None instead of a row list and then ask for index positions.