No id match error handling, csv_merge.py
collection_list() calls collection_from_aip() and had error handling for if an error was return from regex but would catch if there was no match for the group or other condition. Removed places it returned a string so don't have to anticipate places it won't match - instead it will return None if no match and can catch that.