Change order, archive_capacity.py
Now read_usage() outputs a dictionary, the data from collections_count() is added to that dictionary, and once everything is calculated the contents of the dictionary are saved to a csv. Avoids problem of needing to find the correct row in the summary csv to add the collection count to.