import collection_from_aip(), csv_merge_by_aip.py
This is complex enough function that definately don't want to keep two versions in sync. Especially since this will change as new id variations are created.

Not ideal because it runs csv_merge.py and makes that report too, but it does run without errors and that other report is fast so acceptable for now.