archive_overview(), use dataframes with usage, reports.py
Initially read usage report as CSV because do so much data manipulation and dealing with rows that have different kinds of data. But once done with that, convert to a dataframe so it is easy to merge with the other data.

The code as is produces a correct dataframe, just not formatted exactly as I would like.