Reduce files in format csv, merge_format_reports.py
The by format csv is only used for file count subtotals. Having long things like a collection list and especially the aip list causes problems opening it in Excel AND not using that information. My expectation is if someone wants to review data they will use the by aip one or just look at their group csv, so not necessary to have the added stuff if by format.