Do not add totals to type and name, reports.py
Previously added the collection, aip, and file totals to the dataframes for format type and format standardized name. However, when sorting results in the spreadsheet, the total row was getting in the way. Don't need on this sheets since there is a overview sheet with those totals in the spreadsheet.