Add totals parameter to one_category(), reports.py
Instead of relying on global variables for the collection, aip, and file totals, passing to the function as a list.  It was working as global variables but it seemed safer to have a clear indiciation of where the variable values were coming from.

Could have done each total as a separate variable, but it felt simpler to just have the one line to make the variable (instead of 3) and one extra parameter (instead of 3).