Changed filters in count_ranges(), reports.py
I though I might be missing out on the edge number (10, 100, etc.) and adjusted the lower bound from > 9 to >= 10, etc. Did not change the result.

Also added a test for the number of unique categories in df, in case the function was pulling the wrong format df somehow, but it is correct.