Rename series sooner, group_overlap(), reports.py
Was renaming both columns (which have a name Group to start) after conatenation. So that the real names are associated with them sooner, rename as soon as make the series. Otherwise, since have the same original name, have to rely on the order to get the new names with the right columns. This feels more explicit and safe.