Add format id to merged csvs
Was making the id (name|version|key) in reports.py in the dataframe, but decided it would be nice to have in the merged format CSVs instead. Now make it in merge_format_reports.py.

Also removed _ from column headers - makes it harder to read and isn't necessary to work with the data.