Deleted todo with hypertext version, reports.py
Issue solved when switched to adding the format identification to the merged csvs instead of waiting until the dataframe. There are three format idenfications where the format version is null. It is retained as null when making the format identification in the csv with Python. It was being treated as nothing by the dataframe and not making a format identification.