Unit test for update_standardization.py full script
Starting with a unit test of the full script in preparation for breaking it into more functions to allow for easier testing and simplifying some things. This way, I can detect if I break the whole by changing a piece.

Found and fixed a bug where the script was iterating on the wrong folder for format reports.