Alphabetized functions
Makes it easier to find a function to work on. Was in order of the script (fuzzy since some are called more than once), which requires knowing the script order.

No code within any function was changed, other than moving size_in_tb() out of archive_over() and putting it at the same level.