Fix match_nara_risk to work with NO VALUE if blank
In the accessioning script, if the format id had no version or no PUID, the value would be nan (blank). At this point in this script, blanks are filled with NO VALUE. Impacts how construct name_version (remove NO BLANK at end) and what to filter for to separate format id includes a PUID or not.