Add NARA categories to make_df()
This is the only part of csv_to_dataframe() cleanup that is not reproduced by how these dataframes are made. The categories are not currently needed for testing this function, but I wanted this to be accurate for future changes after running into an error in different tests from not having NARA be categories.